# 8259380: Correct pretouch chunk size to cap with actual page size