@tschatzl 2021-01-08
Another option is to just set the default chunk size for aarch64 to e.g. 512M and defer searching for the "best" later.